pyproject.toml
[project]
name = "pilcd0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Yoichi Tanibayashi", email = "yoichi@tanibayashi.jp" }
]
requires-python = ">=3.11"
dependencies = [
    "pigpio>=1.78",
    "pillow>=11.3.0",
    "spidev>=3.7",
]

[build-system]
requires = ["uv_build>=0.8.12,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pylint>=3.3.8",
    "pyright>=1.1.403",
    "pytest>=8.4.1",
    "ruff>=0.12.9",
]
